+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Added a `NEWS.md` file to track changes to the package.  
 * Added `log_*` functions (https://github.com/meerapatelmd/pg13/issues/13)  
-* Added `migrate` function (https://github.com/meerapatelmd/pg13/issues/18)
+* Added `migrate` function (https://github.com/meerapatelmd/pg13/issues/18) 
+* Added `kill` and `getConnDB` functions  
+
+